feat: move format_duration_secs to utils and add new unit tests
